requirements.txt
streamlit
huggingface-hub
python-dotenv
pandas
crewai
crewai-tools
langchain-google-genai
google-generativeai
protobuf==4.25.5
